Modification to the Super Class Mage_Core_Controller_Varien_Action
This change tells Magento to not fail a form key check if the session is
a brand new one. Since the session is brand new, there should be no fear
of a CSFR attack.

additionally, it leaves CSFR testing on for subsequent transactions as
well as leaving CSFR on for the Magento Admin interface.